+++ buildroot/bin/parallel-examples/fix_host_menu_configs.py
@@ -1,27 +1,17 @@
 #!/usr/bin/env python3
 """
-build_and_log.py — Simple parallel build runner with shared error logging.
-
-This is the baseline script for a simple parallel build: it builds each
-config example and logs any failures to a shared log file so problems can
-be triaged after all parallel builds have completed.
+fix_host_menu_configs.py — Fix HOST_* menu-item config options.
 
 Usage:
-    python3 build_and_log.py [--loop-limit=N] [--resume=DIR] [--limit=N]
-                              [--nofail] [--many] [--archive] [-a]
+    python3 fix_host_menu_configs.py [--loop-limit=N] [--resume=DIR] [--limit=N]
 
-The log file is written to .pio/build_errors.log inside MARLIN_REPO,
-so all parallel clones write to the same shared location when their
-MARLIN_REPO env var points at the same place (or can be merged later).
+Phases per config:
+    start        -> build as-is; on failure enable HOST_* menu items, retry
+    continue     -> HOST_* enabled; on failure check prog size or skip
+    skip         -> skip to next config
 """
 
-import subprocess
-import sys
-import os
-import re
-import argparse
-import signal
-import datetime
+import subprocess, sys, os, re, argparse, signal
 
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 PLATFORMIO_SCRIPTS = os.path.join(SCRIPT_DIR, '..', 'share', 'PlatformIO', 'scripts')
@@ -30,15 +20,12 @@ import config  # type: ignore[import-untyped]
 
 # ---------------------------------------------------------------------------
 # Environment — derive repo path from CWD (not a hardcoded env var)
+# This allows the parallel orchestrator to cd into a clone first.
 # ---------------------------------------------------------------------------
 MARLIN_REPO  = os.getcwd()
 MARLIN_CONFIGS = os.environ.get('MARLIN_CONFIGS', '')
 BUILDALL     = os.path.join(MARLIN_REPO, 'buildroot', 'bin', 'build_all_examples')
 STAT_FILE    = os.path.join(MARLIN_REPO, '.pio', '.buildall')
-LOG_FILE     = os.path.join(MARLIN_REPO, '.pio', 'build_errors.log')
-
-# How many lines to capture from the tail of failed build output
-TAIL_LINES = 40
 
 
 def run_build(build_opts=None, resume=None):
@@ -56,7 +43,6 @@ def run_build(build_opts=None, resume=None):
         cwd=MARLIN_REPO,
     )
     output = proc.stdout + proc.stderr
-    # Strip ANSI escape codes so log file is clean
     output_clean = re.sub(r'\x1b\[[0-9;]*m', '', output)
     return proc.returncode, output_clean
 
@@ -75,6 +61,14 @@ def get_failing_config(output):
             pass
 
     if not failing:
+        m = re.search(
+            r'Set\s+HOST_START_MENU_ITEM\s+HOST_SHUTDOWN_MENU_ITEM\s+in:\s+(\S+)',
+            output,
+        )
+        if m:
+            failing = os.path.basename(m.group(1))
+
+    if not failing:
         m = re.search(r'Failed to build\s+(\S+)', output)
         if m:
             failing = m.group(1)
@@ -82,24 +76,22 @@ def get_failing_config(output):
     return failing
 
 
-def ensure_log_dir():
-    """Make sure the .pio directory and log file parent exist."""
-    os.makedirs(os.path.dirname(LOG_FILE), exist_ok=True)
+def enable_config_options(config_dir, *options):
+    """Enable config options in all existing config files under config_dir."""
+    files = config.resolve(config_dir)
+    for fpath in files:
+        if os.path.exists(fpath):
+            for opt in options:
+                config.enable(fpath, opt, True)
 
 
-def log_error(failing, output_tail, extra=""):
-    """Append failure info to the shared error log file."""
-    ensure_log_dir()
-    timestamp = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
-    with open(LOG_FILE, 'a', encoding='utf-8') as f:
-        f.write(f"\n{'='*72}\n")
-        f.write(f"[{timestamp}] FAILURE: {failing}\n")
-        if extra:
-            f.write(f"  Note: {extra}\n")
-        f.write(f"{'='*72}\n")
-        f.write(output_tail)
-        f.write("\n")
-    print(f"  -> Logged to {LOG_FILE}")
+def disable_config_options(config_dir, *options):
+    """Disable config options in all existing config files under config_dir."""
+    files = config.resolve(config_dir)
+    for fpath in files:
+        if os.path.exists(fpath):
+            for opt in options:
+                config.enable(fpath, opt, False)
 
 
 def stash_changes():
@@ -122,7 +114,7 @@ def restore_stash(stashed):
 
 def main():
     parser = argparse.ArgumentParser(
-        description='Build all config examples and log failures to a shared log file.'
+        description='Fix HOST_* menu-item options across all example configs.'
     )
     parser.add_argument(
         '--loop-limit', type=int, default=0,
@@ -132,29 +124,12 @@ def main():
         '--resume', type=str, default=None,
         help='Start at this config directory name',
     )
-    parser.add_argument(
-        '--limit', type=int, default=None,
-        help='Pass --limit to build_all_examples',
-    )
-    parser.add_argument(
-        '--nofail', action='store_true',
-        help='Pass --nofail to build_all_examples (continue on fail)',
-    )
-    parser.add_argument(
-        '--many', action='store_true',
-        help='Pass --many to build_all_examples (all envs per config)',
-    )
-    parser.add_argument(
-        '--archive', action='store_true',
-        help='Pass -a to build_all_examples',
-    )
-    parser.add_argument(
-        '--tail', type=int, default=TAIL_LINES,
-        help=f'Lines of output to log on failure (default: {TAIL_LINES})',
-    )
+    parser.add_argument('--limit', type=int, default=None)
+    parser.add_argument('--nofail', action='store_true')
+    parser.add_argument('--many', action='store_true')
+    parser.add_argument('--archive', action='store_true')
     args, unknown = parser.parse_known_args()
 
-    # Collect passthrough flags
     build_opts = []
     if args.limit:
         build_opts.append(f'--limit={args.limit}')
@@ -170,18 +145,6 @@ def main():
     if os.path.exists(STAT_FILE):
         os.remove(STAT_FILE)
 
-    # Ensure fresh log — back up previous if it exists
-    if os.path.exists(LOG_FILE):
-        backup = LOG_FILE + '.prev'
-        if os.path.exists(backup):
-            os.remove(backup)
-        os.rename(LOG_FILE, backup)
-
-    ensure_log_dir()
-    with open(LOG_FILE, 'w', encoding='utf-8') as f:
-        f.write(f"Build error log started at {datetime.datetime.now().isoformat()}\n")
-
-    # Restore stock configs
     subprocess.run(
         [os.path.join(MARLIN_REPO, 'buildroot', 'bin', 'restore_configs')],
         cwd=MARLIN_REPO, check=True, capture_output=True,
@@ -191,51 +154,76 @@ def main():
 
     def cleanup(signum=None, frame=None):
         restore_stash(stashed)
-        sys.exit(0 if signum is None else 128 + (signum or 0))
+        if signum is not None:
+            sys.exit(128 + signum)
 
     signal.signal(signal.SIGINT, cleanup)
     signal.signal(signal.SIGTERM, cleanup)
 
     try:
         loop_count = 0
+        next_phase = 'start'        # start | continue | skip
+        last_failing = ''
+        patch_count = 0
+        has_prog_size = False
+
         build_args = []
         if args.resume:
             build_args.append(f'--resume={args.resume}')
 
         for iteration in range(1, 501):
-            # Check loop limit
             if args.loop_limit > 0 and loop_count >= args.loop_limit:
                 print(f"Loop limit ({args.loop_limit}) reached.")
                 sys.exit(0)
 
-            # --- Build ---------------------------------------------------
+            # Include build_opts in every invocation
             build_cmd_args = list(build_args) + build_opts
 
-            print(f"=== Iteration {iteration} {' '.join(build_cmd_args)} ===")
+            if next_phase == 'continue':
+                # Remove --resume: --continue takes over from the stat file
+                build_cmd_args = [a for a in build_cmd_args if not a.startswith('--resume=')]
+                build_cmd_args.append('--continue')
+            elif next_phase == 'skip':
+                build_cmd_args = [a for a in build_cmd_args if not a.startswith('--resume=')]
+                build_cmd_args.append('--skip')
+
+            print(f"=== Iteration {iteration}: phase={next_phase} {' '.join(build_cmd_args)} ===")
 
             rc, output = run_build(build_cmd_args)
             print(output)
 
-            # --- Success: this config built fine ---
             if rc == 0:
-                print("Build succeeded for this config.")
+                print(f"ALL DONE — built successfully after {patch_count} patch(es)")
                 sys.exit(0)
 
             loop_count += 1
-
-            # --- Failure: log and move on ---
             failing = get_failing_config(output)
-            if not failing:
-                failing = f"unknown-config-{iteration}"
+            has_prog_size = "Error: The program size" in output
 
-            output_tail = "\n".join(output.splitlines()[-args.tail:])
-            log_error(failing, output_tail)
+            if failing:
+                print(f"HOST_* menu-item error in: {failing}")
+                config_dir = os.path.join(MARLIN_CONFIGS, 'config', 'examples', failing)
+                print(f"Patching {config_dir} ...")
+                enable_config_options(config_dir, 'HOST_START_MENU_ITEM', 'HOST_SHUTDOWN_MENU_ITEM')
+                last_failing = failing
+                patch_count += 1
+                next_phase = 'continue'
 
-            # Clear stat file for next build
+            elif has_prog_size:
+                print(f"Program size overflow — reverting HOST_* patches and skipping {last_failing}.")
+                config_dir = os.path.join(MARLIN_CONFIGS, 'config', 'examples', last_failing)
+                disable_config_options(config_dir, 'HOST_START_MENU_ITEM', 'HOST_SHUTDOWN_MENU_ITEM')
+                next_phase = 'skip'
+
+            else:
+                print("Non-HOST_* error detected — skipping this config.")
+                print("=== Last 30 lines of output ===")
+                for line in output.splitlines()[-30:]:
+                    print(line)
+                next_phase = 'skip'
+
             if os.path.exists(STAT_FILE):
                 os.remove(STAT_FILE)
-
-            print(f"Logged error for {failing}. Moving to next config.")
             print("")
 
         print("Hit max iterations (500)")
