+++ buildroot/bin/use_example_configs
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 #
-# use_example_configs [repo:]configpath
+# use_example_configs [branch:]configpath
 #
 # Examples:
 #   use_example_configs
